Fixed splint errors.

+++ src/extended/seqpos_classifier.c
@@ -135,7 +135,8 @@ int gt_seqpos_classifier_position_is_inside_feature(
     bool *end_of_annotation, GtError *err)
 {
   int had_err = 0;
-  GtRange next_specified_ft_range = { GT_UNDEF_LONG, GT_UNDEF_LONG };
+  GtRange next_specified_ft_range = { (unsigned long) GT_UNDEF_LONG,
+                                      (unsigned long) GT_UNDEF_LONG };
   if (i == 0)
   {
     had_err = gt_seqpos_classifier_next_specified_ft(seqpos_classifier,
